src/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.14
charset-normalizer==2.1.1
configparser==5.3.0
Django==4.1.1
django-bootstrap4==22.2
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
dnspython==2.2.1
ez-setup==0.9
idna==3.4
PyJWT==2.4.0
pymongo==3.12.3
PyMySQL==1.0.2
pytz==2022.2.1
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.2.4
urllib3==1.26.12
